# fast/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Context Relay API",
    "description": "Context Relay System Mock API for Frontend Development",
    "version": "0.1.0"
  },
  "paths": {
    "/context/initialize": {
      "post": {
        "tags": [
          "context"
        ],
        "summary": "Initialize Context",
        "description": "Initialize a new context packet.",
        "operationId": "initialize_context_context_initialize_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitializeContextRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InitializeContextResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/context/relay": {
      "post": {
        "tags": [
          "context"
        ],
        "summary": "Relay Context",
        "description": "Relay context fragments between agents.",
        "operationId": "relay_context_context_relay_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Request"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/context/merge": {
      "post": {
        "tags": [
          "context"
        ],
        "summary": "Merge Contexts",
        "description": "Merge multiple contexts.",
        "operationId": "merge_contexts_context_merge_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Request"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/context/prune": {
      "post": {
        "tags": [
          "context"
        ],
        "summary": "Prune Context",
        "description": "Prune context fragments based on strategy.",
        "operationId": "prune_context_context_prune_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Request"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/context/{context_id}": {
      "get": {
        "tags": [
          "context"
        ],
        "summary": "Get Context",
        "description": "Get a specific context.",
        "operationId": "get_context_context__context_id__get",
        "parameters": [
          {
            "name": "context_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Context Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetContextResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/context/{context_id}/versions": {
      "get": {
        "tags": [
          "context"
        ],
        "summary": "List Versions",
        "description": "List all versions of a context.",
        "operationId": "list_versions_context__context_id__versions_get",
        "parameters": [
          {
            "name": "context_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Context Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListVersionsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgentType": {
        "type": "string",
        "enum": [
          "ai_assistant",
          "human",
          "system",
          "external_service"
        ],
        "title": "AgentType",
        "description": "Types of agents in the context relay system."
      },
...
      "ContextFragment-Output": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Unique fragment identifier"
          },
          "type": {
            "$ref": "#/components/schemas/FragmentType",
            "description": "Type of fragment content"
          },
          "content": {
            "type": "string",
            "title": "Content",
            "description": "Fragment content"
          },
          "source_agent": {
            "$ref": "#/components/schemas/AgentType",
            "description": "Agent that created this fragment"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp",
            "description": "Creation timestamp"
          },
          "embedding": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embedding",
            "description": "Vector embedding for semantic similarity"
          },
          "importance_score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Importance Score",
            "description": "Importance score for pruning",
            "default": 1.0
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata",
            "description": "Additional fragment metadata"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags",
            "description": "Fragment tags for categorization"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "type",
          "content",
          "source_agent"
        ],
        "title": "ContextFragment",
        "description": "Individual context piece with metadata."
      },
...
      "ContextPacket-Output": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Unique context identifier"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id",
            "description": "Session identifier"
          },
          "fragments": {
            "items": {
              "$ref": "#/components/schemas/ContextFragment-Output"
            },
            "type": "array",
            "title": "Fragments",
            "description": "Context fragments"
          },
          "version": {
            "type": "integer",
            "title": "Version",
            "description": "Context version number",
            "default": 0
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Creation timestamp"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "Last update timestamp"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata",
            "description": "Additional context metadata"
          },
          "decision_trace": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Decision Trace",
            "description": "Decision trace entries"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "session_id"
        ],
        "title": "ContextPacket",
        "description": "Core context container with fragments."
      },
      "FragmentType": {
        "type": "string",
        "enum": [
          "text",
          "code",
          "data",
          "metadata",
          "decision"
        ],
        "title": "FragmentType",
        "description": "Types of context fragments."
      },
      "GetContextResponse": {
        "properties": {
          "context": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContextPacket-Output"
              },
              {
                "type": "null"
              }
            ],
            "description": "Requested context"
          },
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Operation success status"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Response message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "title": "GetContextResponse",
        "description": "Response for getting a specific context."
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "InitializeContextRequest": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id",
            "description": "Session identifier"
          },
          "initial_input": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Initial Input",
            "description": "Initial input as string or structured object"
          },
          "initial_fragments": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/app__models__context__ContextFragment"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Initial Fragments",
            "description": "Initial context fragments"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata",
            "description": "Additional metadata"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "session_id"
        ],
        "title": "InitializeContextRequest",
        "description": "Request to initialize a new context.\n\nTest-suite compatibility: supports either `initial_input` (str|object)\nor `initial_fragments` (List[ContextFragment])."
      },
      "InitializeContextResponse": {
        "properties": {
          "context": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ContextPacket-Output"
              },
              {
                "type": "null"
              }
            ],
            "description": "Created context packet"
          },
          "context_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Id",
            "description": "Created context ID"
          },
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Operation success status",
            "default": true
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Response message",
            "default": "Context initialized successfully"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "title": "InitializeContextResponse",
        "description": "Response for context initialization.\n\nFor compatibility, the actual endpoint returns a dict with `context_id`\nand `context_packet`. We keep this model for OpenAPI completeness but do\nnot enforce it at runtime."
      },
      "ListVersionsResponse": {
        "properties": {
          "versions": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Versions",
            "description": "Available versions"
          },
          "current_version": {
            "type": "integer",
            "title": "Current Version",
            "description": "Current version number"
          },
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "Operation success status"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Response message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "current_version",
          "success",
          "message"
        ],
        "title": "ListVersionsResponse",
        "description": "Response for listing context versions."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "app__models__context__ContextFragment": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Unique fragment identifier"
          },
          "type": {
            "$ref": "#/components/schemas/FragmentType",
            "description": "Type of fragment content"
          },
          "content": {
            "type": "string",
            "title": "Content",
            "description": "Fragment content"
          },
          "source_agent": {
            "$ref": "#/components/schemas/AgentType",
            "description": "Agent that created this fragment"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp",
            "description": "Creation timestamp"
          },
          "embedding": {
            "anyOf": [
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embedding",
            "description": "Vector embedding for semantic similarity"
          },
          "importance_score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Importance Score",
            "description": "Importance score for pruning",
            "default": 1.0
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata",
            "description": "Additional fragment metadata"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags",
            "description": "Fragment tags for categorization"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "type",
          "content",
          "source_agent"
        ],
        "title": "ContextFragment",
        "description": "Enhanced context fragment with MongoDB support."
      },
...
    }
  }
}